+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- methods ip_address and ip_details into Mautic::Submissions::Form
+
 ## [2.6.0] - 2021-02-08
 ### Added
 - \#23 Stage